- update to 20090702
Changed files:
    dokuwiki-find-lang.sh -> 1.7
    dokuwiki-plugin-odt.spec -> 1.8

+++ dokuwiki-plugin-odt.spec
@@ -2,21 +2,23 @@
 Summary:	DokuWiki odt (Open Document Text) Export Plugin
 Summary(pl-UTF8):	Wtyczka do eksportowania plików odt (Open Document Text)
 Name:		dokuwiki-plugin-%{plugin}
-Version:	20090115
+Version:	20090702
 Release:	1
 License:	GPL v2
 Group:		Applications/WWW
-Source0:	http://gauret.free.fr/fichiers/dokuwiki/dokuwiki-odt-%{version}.tgz
-# Source0-md5:	278748da7659129ae9806b9726961e53
-Source1:	dokuwiki-find-lang.sh
+Source0:	http://gauret.free.fr/fichiers/dokuwiki/dokuwiki-odt-%{version}.zip
+# Source0-md5:	41c47719f7a118a0ffbc1428c841bb56
 URL:		http://wiki.splitbrain.org/plugin:odt
+Patch0:		dokuwiki-ziplib.patch
 Requires:	dokuwiki >= 20070626
+Requires:	php-common >= 4:5.0
 BuildArch:	noarch
 BuildRoot:	%{tmpdir}/%{name}-%{version}-root-%(id -u -n)
 
 %define		dokuconf	/etc/webapps/dokuwiki
 %define		dokudir		/usr/share/dokuwiki
 %define		plugindir	%{dokudir}/lib/plugins/%{plugin}
+%define		find_lang 	%{_usrlibrpm}/dokuwiki-find-lang.sh %{buildroot}
 
 %description
 This plugin allows you to export a page to the OpenDocument format
@@ -27,22 +29,29 @@ Hint: Open Office can also export to PDF.
 
 %description -l pl-UTF8
 Ta wtyczka pozwala na eksportowanie strony do formatu OpenDocument,
-używanego przez Open Office i inne edytory tekstu. Przydatna jest
-w sytuacji w której musisz dać pojedyńczą stronę klientowi.
+używanego przez Open Office i inne edytory tekstu. Przydatna jest w
+sytuacji w której musisz dać pojedyńczą stronę klientowi.
 
 Podpowiedź: Open Office pozwala także na eksportowanie do PDF.
 
 %prep
 %setup -q -n %{plugin}
+%patch0 -p1
+version=$( awk '/^date/{print $2}' info.txt)
+if [ "$(echo "$version" | tr -d -)" != %{version} ]; then
+	: %%{version} mismatch
+	exit 1
+fi
+
+rm -f ZipLib.class.php
 
 %install
 rm -rf $RPM_BUILD_ROOT
 install -d $RPM_BUILD_ROOT%{plugindir}
-cp -a . $RPM_BUILD_ROOT%{plugindir}
+cp -a conf lang *.php *.png *.xml info.txt $RPM_BUILD_ROOT%{plugindir}
 rm -f $RPM_BUILD_ROOT%{plugindir}/{ChangeLog,.gitignore}
 
-# find locales
-sh %{SOURCE1} %{name}.lang
+%find_lang %{name}.lang
 
 %clean
 rm -rf $RPM_BUILD_ROOT
@@ -54,7 +63,8 @@ if [ -f %{dokuconf}/local.php ]; then
 fi
 
 %files -f %{name}.lang
-%doc ChangeLog
+%defattr(644,root,root,755)
+%doc ChangeLog README.txt TODO.txt
 %defattr(644,root,root,755)
 %dir %{plugindir}
 %{plugindir}/*.txt
